refactor(mcp): mechanically split tool_call_handler into bounded modules (wave16 step2)
