+++ crates/assay-core/src/mcp/tool_call_handler/evaluate.rs
@@ -1,0 +1,381 @@
+use super::super::decision::{reason_codes, DecisionEmitterGuard};
+use super::super::identity::ToolIdentity;
+use super::super::jsonrpc::JsonRpcRequest;
+use super::super::lifecycle::mandate_used_event;
+use super::super::policy::{PolicyDecision, PolicyState};
+use super::emit;
+use super::types::{HandleResult, ToolCallHandler};
+use crate::runtime::{AuthorizeError, MandateData, OperationClass, ToolCallData};
+use serde_json::Value;
+use std::time::Instant;
+
+pub(super) fn handle_tool_call(
+    handler: &ToolCallHandler,
+    request: &JsonRpcRequest,
+    state: &mut PolicyState,
+    runtime_identity: Option<&ToolIdentity>,
+    mandate: Option<&MandateData>,
+    transaction_object: Option<&Value>,
+) -> HandleResult {
+    let params = match request.tool_params() {
+        Some(p) => p,
+        None => {
+            // Not a tool call - still must emit decision (I1 invariant)
+            let tool_call_id = handler.extract_tool_call_id(request);
+            let guard = DecisionEmitterGuard::new(
+                handler.emitter.clone(),
+                handler.config.event_source.clone(),
+                tool_call_id.clone(),
+                "unknown".to_string(),
+            );
+            guard.emit_error(
+                reason_codes::S_INTERNAL_ERROR,
+                Some("Not a tool call".to_string()),
+            );
+
+            return emit::error_not_tool_call(&handler.config.event_source, tool_call_id);
+        }
+    };
+
+    let tool_name = params.name.clone();
+    let tool_call_id = handler.extract_tool_call_id(request);
+
+    // Create guard - ensures decision is ALWAYS emitted
+    let mut guard = DecisionEmitterGuard::new(
+        handler.emitter.clone(),
+        handler.config.event_source.clone(),
+        tool_call_id.clone(),
+        tool_name.clone(),
+    );
+    guard.set_request_id(request.id.clone());
+
+    let start = Instant::now();
+
+    // Step 1: Policy evaluation
+    let policy_eval = handler.policy.evaluate_with_metadata(
+        &tool_name,
+        &params.arguments,
+        state,
+        runtime_identity,
+    );
+    let tool_classes = policy_eval.metadata.tool_classes.clone();
+    let matched_tool_classes = policy_eval.metadata.matched_tool_classes.clone();
+    let match_basis = policy_eval
+        .metadata
+        .match_basis
+        .as_str()
+        .map(ToString::to_string);
+    let matched_rule = policy_eval.metadata.matched_rule.clone();
+    guard.set_tool_match(
+        policy_eval.metadata.tool_classes.clone(),
+        policy_eval.metadata.matched_tool_classes.clone(),
+        match_basis.clone(),
+        matched_rule.clone(),
+    );
+
+    match policy_eval.decision {
+        PolicyDecision::Deny {
+            tool: _,
+            code,
+            reason,
+            contract: _,
+        } => {
+            let reason_code = handler.map_policy_code_to_reason(&code);
+            guard.emit_deny(&reason_code, Some(reason.clone()));
+
+            return emit::deny(
+                &handler.config.event_source,
+                tool_call_id,
+                tool_name,
+                &reason_code,
+                reason,
+                emit::ToolMatchMetadata::new(
+                    tool_classes,
+                    matched_tool_classes,
+                    match_basis,
+                    matched_rule,
+                ),
+            );
+        }
+        PolicyDecision::AllowWithWarning { .. } | PolicyDecision::Allow => {
+            // Continue to mandate check
+        }
+    }
+
+    // Step 2: Check if mandate is required
+    let is_commit_tool = handler.is_commit_tool(&tool_name);
+    if is_commit_tool && handler.config.require_mandate_for_commit && mandate.is_none() {
+        let reason = "Commit tool requires mandate authorization".to_string();
+        guard.emit_deny(reason_codes::P_MANDATE_REQUIRED, Some(reason.clone()));
+
+        return emit::deny(
+            &handler.config.event_source,
+            tool_call_id,
+            tool_name,
+            reason_codes::P_MANDATE_REQUIRED,
+            reason,
+            emit::ToolMatchMetadata::new(
+                tool_classes,
+                matched_tool_classes,
+                match_basis,
+                matched_rule,
+            ),
+        );
+    }
+
+    // Step 3: Mandate authorization (if mandate present)
+    if let (Some(authorizer), Some(mandate_data)) = (&handler.authorizer, mandate) {
+        let operation_class = handler.operation_class_for_tool(&tool_name);
+
+        let tool_call_data = ToolCallData {
+            tool_name: tool_name.clone(),
+            tool_call_id: tool_call_id.clone(),
+            operation_class,
+            transaction_object: transaction_object.cloned(),
+            source_run_id: None,
+        };
+
+        let authz_start = Instant::now();
+        match authorizer.authorize_and_consume(mandate_data, &tool_call_data) {
+            Ok(receipt) => {
+                let authz_ms = authz_start.elapsed().as_millis() as u64;
+                guard.set_mandate_info(
+                    Some(mandate_data.mandate_id.clone()),
+                    Some(receipt.use_id.clone()),
+                    Some(receipt.use_count),
+                );
+                guard.set_mandate_matches(Some(true), Some(true), Some(true));
+                guard.set_latencies(Some(authz_ms), None);
+                guard.emit_allow(reason_codes::P_MANDATE_VALID);
+
+                // Emit mandate.used lifecycle event (P0-B)
+                // Only emit on first consumption, not on idempotent retries
+                if receipt.was_new {
+                    if let Some(ref lifecycle) = handler.lifecycle_emitter {
+                        let event = mandate_used_event(&handler.config.event_source, &receipt);
+                        lifecycle.emit(&event);
+                    }
+                }
+
+                return emit::allow(
+                    &handler.config.event_source,
+                    tool_call_id,
+                    tool_name,
+                    reason_codes::P_MANDATE_VALID,
+                    Some(receipt),
+                    emit::ToolMatchMetadata::new(
+                        tool_classes,
+                        matched_tool_classes,
+                        match_basis,
+                        matched_rule,
+                    ),
+                );
+            }
+            Err(e) => {
+                let (reason_code, reason) = handler.map_authz_error(&e);
+                guard.set_mandate_info(Some(mandate_data.mandate_id.clone()), None, None);
+                guard.emit_deny(&reason_code, Some(reason.clone()));
+
+                return emit::deny(
+                    &handler.config.event_source,
+                    tool_call_id,
+                    tool_name,
+                    &reason_code,
+                    reason,
+                    emit::ToolMatchMetadata::new(
+                        tool_classes,
+                        matched_tool_classes,
+                        match_basis,
+                        matched_rule,
+                    ),
+                );
+            }
+        }
+    }
+
+    // Step 4: No mandate required, policy allows
+    let elapsed_ms = start.elapsed().as_millis() as u64;
+    guard.set_latencies(Some(elapsed_ms), None);
+    guard.emit_allow(reason_codes::P_POLICY_ALLOW);
+
+    emit::allow(
+        &handler.config.event_source,
+        tool_call_id,
+        tool_name,
+        reason_codes::P_POLICY_ALLOW,
+        None,
+        emit::ToolMatchMetadata::new(
+            tool_classes,
+            matched_tool_classes,
+            match_basis,
+            matched_rule,
+        ),
+    )
+}
+
+impl ToolCallHandler {
+    /// Extract tool_call_id from request (I4: idempotency key).
+    pub(super) fn extract_tool_call_id(&self, request: &JsonRpcRequest) -> String {
+        // Try to get from params._meta.tool_call_id (MCP standard)
+        if let Some(params) = request.tool_params() {
+            if let Some(meta) = params.arguments.get("_meta") {
+                if let Some(id) = meta.get("tool_call_id").and_then(|v| v.as_str()) {
+                    return id.to_string();
+                }
+            }
+        }
+
+        // Fall back to request.id if present
+        if let Some(id) = &request.id {
+            if let Some(s) = id.as_str() {
+                return format!("req_{}", s);
+            }
+            if let Some(n) = id.as_i64() {
+                return format!("req_{}", n);
+            }
+        }
+
+        // Generate one if none found
+        format!("gen_{}", uuid::Uuid::new_v4())
+    }
+
+    /// Check if a tool is classified as a commit operation.
+    pub(super) fn is_commit_tool(&self, tool_name: &str) -> bool {
+        self.config.commit_tools.iter().any(|pattern| {
+            if pattern == "*" {
+                return true;
+            }
+            if pattern.ends_with('*') {
+                let prefix = pattern.trim_end_matches('*');
+                tool_name.starts_with(prefix)
+            } else {
+                tool_name == pattern
+            }
+        })
+    }
+
+    /// Check if a tool is classified as a write operation (non-commit).
+    fn is_write_tool(&self, tool_name: &str) -> bool {
+        self.config.write_tools.iter().any(|pattern| {
+            if pattern == "*" {
+                return true;
+            }
+            if pattern.ends_with('*') {
+                let prefix = pattern.trim_end_matches('*');
+                tool_name.starts_with(prefix)
+            } else {
+                tool_name == pattern
+            }
+        })
+    }
+
+    /// Derive operation class from tool classification (commit_tools, write_tools, else Read).
+    pub(super) fn operation_class_for_tool(&self, tool_name: &str) -> OperationClass {
+        if self.is_commit_tool(tool_name) {
+            OperationClass::Commit
+        } else if self.is_write_tool(tool_name) {
+            OperationClass::Write
+        } else {
+            OperationClass::Read
+        }
+    }
+
+    /// Map policy error code to reason code.
+    pub(super) fn map_policy_code_to_reason(&self, code: &str) -> String {
+        match code {
+            "E_TOOL_DENIED" => reason_codes::P_TOOL_DENIED.to_string(),
+            "E_TOOL_NOT_ALLOWED" => reason_codes::P_TOOL_NOT_ALLOWED.to_string(),
+            "E_ARG_SCHEMA" => reason_codes::P_ARG_SCHEMA.to_string(),
+            "E_RATE_LIMIT" => reason_codes::P_RATE_LIMIT.to_string(),
+            "E_TOOL_DRIFT" => reason_codes::P_TOOL_DRIFT.to_string(),
+            _ => reason_codes::P_POLICY_DENY.to_string(),
+        }
+    }
+
+    /// Map authorization error to reason code and message.
+    pub(super) fn map_authz_error(&self, error: &AuthorizeError) -> (String, String) {
+        match error {
+            AuthorizeError::Policy(pe) => {
+                use crate::runtime::PolicyError;
+                match pe {
+                    PolicyError::Expired { .. } => (
+                        reason_codes::M_EXPIRED.to_string(),
+                        "Mandate expired".to_string(),
+                    ),
+                    PolicyError::NotYetValid { .. } => (
+                        reason_codes::M_NOT_YET_VALID.to_string(),
+                        "Mandate not yet valid".to_string(),
+                    ),
+                    PolicyError::ToolNotInScope { tool } => (
+                        reason_codes::M_TOOL_NOT_IN_SCOPE.to_string(),
+                        format!("Tool '{}' not in mandate scope", tool),
+                    ),
+                    PolicyError::KindMismatch { kind, op_class } => (
+                        reason_codes::M_KIND_MISMATCH.to_string(),
+                        format!(
+                            "Mandate kind '{}' does not allow operation class '{}'",
+                            kind, op_class
+                        ),
+                    ),
+                    PolicyError::AudienceMismatch { expected, actual } => (
+                        reason_codes::M_AUDIENCE_MISMATCH.to_string(),
+                        format!(
+                            "Audience mismatch: expected '{}', got '{}'",
+                            expected, actual
+                        ),
+                    ),
+                    PolicyError::IssuerNotTrusted { issuer } => (
+                        reason_codes::M_ISSUER_NOT_TRUSTED.to_string(),
+                        format!("Issuer '{}' not in trusted list", issuer),
+                    ),
+                    PolicyError::MissingTransactionObject => (
+                        reason_codes::M_TRANSACTION_REF_MISMATCH.to_string(),
+                        "Transaction object required but not provided".to_string(),
+                    ),
+                    PolicyError::TransactionRefMismatch { expected, actual } => (
+                        reason_codes::M_TRANSACTION_REF_MISMATCH.to_string(),
+                        format!(
+                            "Transaction ref mismatch: expected '{}', computed '{}'",
+                            expected, actual
+                        ),
+                    ),
+                }
+            }
+            AuthorizeError::Store(se) => {
+                use crate::runtime::AuthzError;
+                match se {
+                    AuthzError::AlreadyUsed => (
+                        reason_codes::M_ALREADY_USED.to_string(),
+                        "Single-use mandate already consumed".to_string(),
+                    ),
+                    AuthzError::MaxUsesExceeded { max, current } => (
+                        reason_codes::M_MAX_USES_EXCEEDED.to_string(),
+                        format!("Max uses exceeded: {} of {} used", current, max),
+                    ),
+                    AuthzError::NonceReplay { nonce } => (
+                        reason_codes::M_NONCE_REPLAY.to_string(),
+                        format!("Nonce replay detected: {}", nonce),
+                    ),
+                    AuthzError::MandateNotFound { mandate_id } => (
+                        reason_codes::M_NOT_FOUND.to_string(),
+                        format!("Mandate not found: {}", mandate_id),
+                    ),
+                    AuthzError::Revoked { revoked_at } => (
+                        reason_codes::M_REVOKED.to_string(),
+                        format!("Mandate revoked at {}", revoked_at),
+                    ),
+                    AuthzError::MandateConflict { .. }
+                    | AuthzError::InvalidConstraints { .. }
+                    | AuthzError::Database(_) => (
+                        reason_codes::S_DB_ERROR.to_string(),
+                        format!("Database error: {}", se),
+                    ),
+                }
+            }
+            AuthorizeError::TransactionRef(msg) => (
+                reason_codes::M_TRANSACTION_REF_MISMATCH.to_string(),
+                format!("Transaction ref error: {}", msg),
+            ),
+        }
+    }
+}
